multi_grid: Move HEALPix grid to grid_impl
